in progress

+++ TeamEdge/WebLayer/Controllers/PartisipantController.cs
@@ -1,38 +1,50 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using System.Linq;
 using System.Threading.Tasks;
 using TeamEdge.BusinessLogicLayer.Interfaces;
-using TeamEdge.DAL.Context;
-using TeamEdge.DAL.Models;
 using TeamEdge.Models;
 
 namespace TeamEdge.WebLayer.Controllers
 {
     [Authorize]
+    [Route("api/partisipant")]
     public class PartisipantController : Controller
     {
         readonly IMembershipService _membershipService;
-        readonly TeamEdgeDbContext _context;
+        IEmailService _emailService;
         readonly IMapper _mapper;
 
-        public PartisipantController(IMembershipService membershipService, IMapper mapper)
+        public PartisipantController(IMembershipService membershipService, IMapper mapper, IEmailService emailService)
         {
             _membershipService = membershipService;
             _mapper = mapper;
+            _emailService = emailService;
         }
 
-        public async Task<IActionResult> SendInvite(CreateInviteVM model)
+        /// <summary>
+        /// Create invite for new project partisipant
+        /// </summary>
+        /// <param name="model"></param>
+        /// <returns></returns>
+        [HttpPost("invite")]
+        public async Task<IActionResult> SendInvite([FromBody]CreateInviteVM model)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var dto = _mapper.Map<CreateInviteDTO>(model);
             dto.FromUserId = User.Id();
-            var operRes = await _membershipService.SendInvite(dto);
-            return operRes.GetResult();
+            var result = await _membershipService.CreateInvite(dto);
+            await _emailService.SendInviteAsync(User.Model(), result);
+            return Ok(result);
         }
 
+        /// <summary>
+        /// Accept invite and join to project
+        /// </summary>
+        /// <param name="inviteId"></param>
+        /// <returns></returns>
+        [HttpPost("join/{inviteId}")]
         public async Task<IActionResult> JoinProject(int inviteId)
         {
             var operRes = await _membershipService.JoinProject(new JoinProjectDTO {
@@ -41,27 +53,51 @@ namespace TeamEdge.WebLayer.Controllers
                 Email = User.Email()});
             return operRes.GetResult();
         }
-
+        
+        /// <summary>
+        /// Remove participant from project
+        /// </summary>
+        /// <param name="projectId"></param>
+        /// <param name="userId"></param>
+        /// <returns></returns>
+        [HttpDelete("project/{projectId}/user/{userId}")]
         public async Task<IActionResult> LeaveProject(int projectId, int userId)
         {
-            var result = await _membershipService.DeletePartisipant(new DeletePartisipantDTO()
-            {
-                FromId = User.Id(),
-                ProjectId = projectId,
-                UserId = userId
-            });
+            var selfLeaving = User.Id() == userId;
 
-            return result.GetResult();
+            if (selfLeaving)
+            {
+                var result = await _membershipService.LeaveProject(userId, projectId);
+                if (result.Succeded)
+                    _emailService.NotifyUserLeaveAsync(User.Model(), projectId);
+                return result.GetResult();
+            }
+            else
+            {
+                await _membershipService.DeletePartisipant(new DeletePartisipantDTO()
+                {
+                    FromId = User.Id(),
+                    ProjectId = projectId,
+                    UserId = userId
+                });
+                _emailService.NotifyUserDeletedAsync(User.Model(), userId, projectId);
+                return Ok();
+            }
         }
 
+
+        /// <summary>
+        /// Change partisipant status
+        /// </summary>
+        /// <param name="model"></param>
+        /// <returns></returns>
+        [HttpPut("status")]
         public async Task<IActionResult> ChangePartisipantStatus(ChangeStatusVM model)
         {
             var dto = _mapper.Map<ChangeStatusDTO>(model);
             dto.UserId = User.Id();
-
-            var result = await _membershipService.UpdatePartisipantStatus(dto);
-
-            return result.GetResult();
+            await _membershipService.UpdatePartisipantStatus(dto);
+            return Ok();
         }
     }
 }
